+++ FrameLib_Framework/FrameLib_Memory.cpp
@@ -285,11 +285,11 @@ FrameLib_LocalAllocator::Storage::~Storage()
     mAllocator.dealloc(mData);
 }
 
-void FrameLib_LocalAllocator::Storage::resize(bool tagged, size_t size)
+void FrameLib_LocalAllocator::Storage::resize(bool tagged, unsigned long size)
 {
     size_t actualSize = tagged ? Serial::inPlaceSize(size) : size * sizeof(double);
-    unsigned long maxSize = actualSize << 1;
-    
+    size_t maxSize = actualSize << 1;
+
     if (mMaxSize >= maxSize)
     {
         // Reallocate for tagged frames
