+++ FrameLib_Objects/IO/FrameLib_Source.h
@@ -32,8 +32,8 @@ private:
     
     // Helpers
     
-    size_t bufferSize() const { return mBuffer.size(); }
-    
+    unsigned long bufferSize() const { return static_cast<unsigned long>(mBuffer.size()); }
+
     unsigned long convertTimeToSamples(double time);
     
     void copy(const double *input, unsigned long offset, unsigned long size);
