+++ FrameLib_Objects/IO/FrameLib_Trace.h
@@ -32,7 +32,7 @@ private:
     
     // Helpers
     
-    size_t bufferSize() const { return mBuffer.size(); }
+    unsigned long bufferSize() const { return static_cast<unsigned long>(mBuffer.size()); }
 
     unsigned long convertTimeToSamples(double time);
